cleaned up with abstraction of manipulation primitives and config file for constants
